disk server and tools 64-bit clean (except filesystem-specific tools)
git-svn-id: http://opencores.org/ocsvn/eco32/eco32/trunk@258 133310ea-c0b2-4f2d-ae91-eb8ea07367b3

+++ disk/tools/mkmboot/stage2/mboot.c
@@ -23,9 +23,9 @@ unsigned int numSectors = 0;	/* gets loaded by previous stage */
 
 
 typedef struct {
-  unsigned long type;
-  unsigned long start;
-  unsigned long size;
+  unsigned int type;
+  unsigned int start;
+  unsigned int size;
   char descr[DESCR_SIZE];
 } PartEntry;
